refactor: dep cycle
